+++ it/server/src/test/java/com/walmartlabs/concord/it/server/TemplateMergeIT.java
@@ -32,7 +32,7 @@ public class TemplateMergeIT extends AbstractServerIT {
         ProjectResource projectResource = proxy(ProjectResource.class);
         projectResource.createOrUpdate(new ProjectEntry(null, projectName, null, null, null, null,
                 Collections.singletonMap(InternalConstants.Request.TEMPLATE_KEY, template.toUri().toString()),
-                null, null));
+                null, null, true));
 
         // ---