+++ server/impl/src/test/java/com/walmartlabs/concord/server/project/ProjectDaoTest.java
@@ -36,7 +36,7 @@ public class ProjectDaoTest extends AbstractDaoTest {
 
         Map<String, Object> cfg = ImmutableMap.of("a", "a-v");
         String projectName = "project#" + System.currentTimeMillis();
-        UUID projectId = projectDao.insert(orgId, projectName, "test", null, cfg, null);
+        UUID projectId = projectDao.insert(orgId, projectName, "test", null, cfg, null, true);
 
         // ---
         Map<String, Object> actualCfg = projectDao.getConfiguration(projectId);
@@ -86,8 +86,8 @@ public class ProjectDaoTest extends AbstractDaoTest {
         String aName = "aProject#" + System.currentTimeMillis();
         String bName = "bProject#" + System.currentTimeMillis();
 
-        projectDao.insert(orgId, aName, "test", null, null, null);
-        projectDao.insert(orgId, bName, "test", null, null, null);
+        projectDao.insert(orgId, aName, "test", null, null, null, true);
+        projectDao.insert(orgId, bName, "test", null, null, null, true);
 
         // ---