+++ server/impl/src/test/java/com/walmartlabs/concord/server/security/secret/SecretDaoTest.java
@@ -26,7 +26,7 @@ public class SecretDaoTest extends AbstractDaoTest {
         String projectName = "project#" + System.currentTimeMillis();
 
         ProjectDao projectDao = new ProjectDao(getConfiguration());
-        UUID projectId = projectDao.insert(orgId, projectName, "test", null, null, null);
+        UUID projectId = projectDao.insert(orgId, projectName, "test", null, null, null, true);
 
         String secretName = "secret#" + System.currentTimeMillis();
         SecretDao secretDao = new SecretDao(getConfiguration());